Async API is intoduced for Route4MeManager and Route4MeManagerV5.
Mix of sync/async context is removed for async flow.
IDs are switched from 32 bit integer to 64 bit integer (both signed and unsigned).
Async API for FastProcessing (JSON) is introduced.
Unnesessary internal state is removed from FastProcessing.
ArrayList is removed from geocoding API.
Internal sleeps and spin waits are removed.
Naming is updated to follow CamelCase.
Internal code smells are removed.
NewtonJson usage is fixed.

+++ route4me-csharp-sdk/Route4MeSDKLibrary/DataTypes/Address.cs
@@ -20,7 +20,7 @@ namespace Route4MeSDK.DataTypes
         ///     Route destination ID
         /// </summary>
         [DataMember(Name = "route_destination_id", EmitDefaultValue = false)]
-        public int? RouteDestinationId { get; set; }
+        public long? RouteDestinationId { get; set; }
 
         /// <summary>
         ///     Route alias
@@ -163,7 +163,7 @@ namespace Route4MeSDK.DataTypes
         ///     favorite addresses.
         /// </summary>
         [DataMember(Name = "contact_id", EmitDefaultValue = false)]
-        public int? ContactId { get; set; }
+        public long? ContactId { get; set; }
 
         /// <summary>
         ///     The status flag to mark an address as visited (aka check in).
